remove shared namespace from notification
