+++ app/components/notification_component/notification_component.html.slim
