+++ app/components/previews/notification_component_preview.rb
@@ -1,0 +1,21 @@
+class Previews::NotificationComponentPreview < ViewComponent::Preview
+  layout "design_system"
+
+  def self.component_name
+    component_class.to_s.underscore.humanize.split.first.downcase
+  end
+
+  def self.component_class
+    NotificationComponent
+  end
+
+  def self.form
+    Previews::NotificationComponentPreview::OptionsForm
+  end
+
+  def self.interactive_options
+    %w[background dismiss icon]
+  end
+
+  def preview; end
+end
