+++ app/views/jobseekers/job_applications/new.html.slim
@@ -16,7 +16,7 @@
     p.govuk-body = t(".description.closing")
     p.govuk-body = t(".description.save_and_return")
 
-    = render(Shared::NotificationComponent.new(body: t(".privacy_info"), variant: "notice", background: true, dismiss: false, icon: true))
+    = render(NotificationComponent.new(body: t(".privacy_info"), variant: "notice", background: true, dismiss: false, icon: true))
     = govuk_button_to t("buttons.start_application"), jobseekers_job_job_application_path(vacancy.id), class: "govuk-button--start"
 
     = govuk_link_to t("buttons.cancel"), job_path(vacancy)
