+++ app/views/nqt_job_alerts/new.html.slim
@@ -28,7 +28,7 @@
 
         = recaptcha
 
-        = render(Shared::NotificationComponent.new(body: t("nqt_job_alerts.intro.unsubscribe"), variant: "notice", background: true, icon: true, dismiss: false))
+        = render(NotificationComponent.new(body: t("nqt_job_alerts.intro.unsubscribe"), variant: "notice", background: true, icon: true, dismiss: false))
         = f.govuk_submit t("buttons.subscribe"), classes: "nqt-job-alert-subscribe-gtm govuk-!-padding-left-8 govuk-!-padding-right-8"
 
   .govuk-grid-row
