+++ app/views/publishers/vacancies/_error_messages.html.slim
@@ -1,2 +1,2 @@
 - if errors.any?
-  = render(Shared::NotificationComponent.new(title: t("messages.jobs.action_required.heading"), body: t("messages.jobs.action_required.message"), variant: "warning", links: errors.to_hash))
+  = render(NotificationComponent.new(title: t("messages.jobs.action_required.heading"), body: t("messages.jobs.action_required.message"), variant: "warning", links: errors.to_hash))
