+++ app/views/publishers/vacancies/documents/destroy.html.slim
@@ -1,2 +1,2 @@
 - flash.each do |style, message|
-  = render(Shared::NotificationComponent.new(body: message, variant: style))
+  = render(NotificationComponent.new(body: message, variant: style))
