+++ app/views/shared/vacancy/_jobseeker_view.html.slim
@@ -53,7 +53,7 @@
   .govuk-grid-column-two-thirds
 
     - if @vacancy.organisations.many?
-      = render(Shared::NotificationComponent.new(body: t("messages.jobs.multi_school_job_notification_html", organisation_type: organisation_type_basic(@vacancy.parent_organisation)), variant: "notice", background: true, dismiss: false, icon: true))
+      = render(NotificationComponent.new(body: t("messages.jobs.multi_school_job_notification_html", organisation_type: organisation_type_basic(@vacancy.parent_organisation)), variant: "notice", background: true, dismiss: false, icon: true))
     .content-pills class="govuk-!-margin-bottom-6"
       h3.govuk-heading-s class="govuk-!-margin-bottom-3" = t("jobs.contents")
       .govuk-grid-row
